Refactor event handling and testing for Kafka and Postgres integration
- Removed obsolete Kafka test file and integrated Kafka client into event handling.
- Updated produce_exchange_data.go to use the new Kafka client structure.
- Enhanced produce_exchange_data_test.go with SQL mocking for better test coverage.
- Refactored signal_candlestick.go to improve error handling and logging.
- Deleted signal_candlestick_test.go due to redundancy and lack of relevance.
- Introduced new to_pg_candlestick.go for handling candlestick data in Postgres.
- Updated to_pg_orderbook.go to improve error handling and logging.
- Removed deprecated to_pg_agg_trade.go and its test file.
- Adjusted server/http.go to use the new database client structure.

+++ pkg/entities/exchange_test.go
@@ -18,7 +18,7 @@ func TestExchange_FromDto(t *testing.T) {
 
 	assert.Equal(t, req.Name, exchange.Name)
 	assert.Equal(t, req.WsUrl, exchange.WsUrl)
-	assert.Equal(t, ExchangeActive, exchange.IsActive)
+	assert.Equal(t, uint(ExchangeActive), exchange.IsActive)
 }
 
 func TestExchange_ToDto(t *testing.T) {
